Add SQL-backed projection position store
